bot/requirements.txt
aiogram~=2.24
requests~=2.27.1
re==2.2.1

aiohttp~=3.8.3
python-dotenv~=0.21.0